requirements.txt
PySide6>=6.5
Pillow>=10.0
numpy>=1.24
# Audio loading (one of these is needed):
librosa>=0.10          # preferred - handles mp3/wav/ogg/flac
soundfile>=0.12        # fallback for wav/flac/ogg
# ffmpeg must be installed system-wide for export
